+++ src/components/Header/Header.test.tsx
@@ -102,6 +102,136 @@ describe("Test the Header component", () => {
     expect(testOnClick).toHaveBeenCalled();
   });
 
+  it("nav items should expand and collapse on click", () => {
+    const { container } = render(
+      <Header
+        navData={navData}
+        headerLogo={
+          <Link href="" title="Project Title">
+            <Logo altText={"CMS.gov Project"} />
+          </Link>
+        }
+      />
+    );
+    const currentSection = screen.getByText("Current Section");
+    const navSection = container.getElementsByClassName("usa-nav__submenu")[0]; // assumes 1st
+
+    // Assert collapsed
+    expect(navSection).toHaveAttribute("hidden");
+
+    fireEvent.click(currentSection);
+
+    // Assert expanded
+    expect(navSection).not.toHaveAttribute("hidden");
+
+    fireEvent.click(currentSection);
+
+    // Assert collapsed
+    expect(navSection).toHaveAttribute("hidden");
+  });
+
+  it("nav items should collapse on outside click", () => {
+    const { container } = render(
+      <Header
+        navData={navData}
+        headerLogo={
+          <Link href="" title="Project Title">
+            <Logo altText={"CMS.gov Project"} />
+          </Link>
+        }
+      />
+    );
+    const body = document.body;
+    const currentSection = screen.getByText("Current Section");
+    const navSection = container.getElementsByClassName("usa-nav__submenu")[0]; // assumes 1st
+
+    // Assert collapsed
+    expect(navSection).toHaveAttribute("hidden");
+
+    fireEvent.click(currentSection);
+
+    // Assert expanded
+    expect(navSection).not.toHaveAttribute("hidden");
+
+    fireEvent.mouseDown(body);
+
+    // Assert collapsed
+    expect(navSection).toHaveAttribute("hidden");
+  });
+
+  it("should render correcty in mobile view", () => {
+    const { container } = render(
+      <Header
+        navData={navData}
+        headerLogo={
+          <Link href="" title="Project Title">
+            <Logo altText={"CMS.gov Project"} />
+          </Link>
+        }
+      />
+    );
+    const menuBtn = screen.getByText("Menu");
+    const closeBtn = container.getElementsByClassName("usa-nav__close")[0];
+    const body = document.body;
+    const nav = container.getElementsByClassName("usa-nav")[0];
+    const overlay = container.getElementsByClassName("usa-overlay")[0];
+
+    // Assert appropriate default classes
+    expect(body.className).toBe("");
+    expect(nav.className).toBe("usa-nav");
+    expect(overlay.className).toBe("usa-overlay");
+
+    fireEvent.click(menuBtn);
+
+    // Assert "is-visible" classes added
+    expect(body.className).toBe("usa-js-mobile-nav--active");
+    expect(nav.className).toBe("usa-nav is-visible");
+    expect(overlay.className).toBe("usa-overlay is-visible");
+
+    fireEvent.click(closeBtn);
+
+    // Assert "is-visible" classes removed
+    expect(body.className).toBe("");
+    expect(nav.className).toBe("usa-nav");
+    expect(overlay.className).toBe("usa-overlay");
+  });
+
+  it("in mobile view, clicking outside the menu should trigger close", () => {
+    const { container } = render(
+      <Header
+        navData={navData}
+        headerLogo={
+          <Link href="" title="Project Title">
+            <Logo altText={"CMS.gov Project"} />
+          </Link>
+        }
+      />
+    );
+    const menuBtn = screen.getByText("Menu");
+    const body = document.body;
+    const nav = container.getElementsByClassName("usa-nav")[0];
+    const overlay = container.getElementsByClassName("usa-overlay")[0];
+
+    // Assert appropriate default classes
+    expect(body.className).toBe("");
+    expect(nav.className).toBe("usa-nav");
+    expect(overlay.className).toBe("usa-overlay");
+
+    fireEvent.click(menuBtn);
+
+    // Assert "is-visible" classes added
+    expect(body.className).toBe("usa-js-mobile-nav--active");
+    expect(nav.className).toBe("usa-nav is-visible");
+    expect(overlay.className).toBe("usa-overlay is-visible");
+
+    fireEvent.mouseDown(body); // outside click
+
+    // Assert "is-visible" classes removed
+    expect(body.className).toBe("");
+    expect(nav.className).toBe("usa-nav");
+    expect(overlay.className).toBe("usa-overlay");
+  });
+
   describe("current prop should add an underline", () => {
     // current should control the active state of the button
     it("should exist on Current Section", () => {